Remove Component namespace on state
