+++ src/Bundle/DependencyInjection/SyliusResourceExtension.php
@@ -26,10 +26,10 @@ use Sylius\Component\Resource\Metadata\AsResource;
 use Sylius\Component\Resource\Metadata\Metadata;
 use Sylius\Component\Resource\Metadata\ResourceMetadata;
 use Sylius\Component\Resource\Reflection\ClassReflection;
-use Sylius\Component\Resource\State\ProcessorInterface;
-use Sylius\Component\Resource\State\ProviderInterface;
-use Sylius\Component\Resource\State\ResponderInterface;
 use Sylius\Component\Resource\Twig\Context\Factory\ContextFactoryInterface;
+use Sylius\Resource\State\ProcessorInterface;
+use Sylius\Resource\State\ProviderInterface;
+use Sylius\Resource\State\ResponderInterface;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
