+++ src/Component/Symfony/EventListener/FactoryListener.php
@@ -16,8 +16,8 @@ namespace Sylius\Component\Resource\Symfony\EventListener;
 use Sylius\Component\Resource\Metadata\FactoryAwareOperationInterface;
 use Sylius\Component\Resource\Metadata\Operation;
 use Sylius\Component\Resource\Metadata\Operation\HttpOperationInitiatorInterface;
-use Sylius\Component\Resource\State\FactoryInterface;
 use Sylius\Resource\Context\Initiator\RequestContextInitiatorInterface;
+use Sylius\Resource\State\FactoryInterface;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
 
 final class FactoryListener
