+++ src/Component/Symfony/EventListener/RespondListener.php
@@ -14,8 +14,8 @@ declare(strict_types=1);
 namespace Sylius\Component\Resource\Symfony\EventListener;
 
 use Sylius\Component\Resource\Metadata\Operation\HttpOperationInitiatorInterface;
-use Sylius\Component\Resource\State\ResponderInterface;
 use Sylius\Resource\Context\Initiator\RequestContextInitiatorInterface;
+use Sylius\Resource\State\ResponderInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\ViewEvent;
 use Webmozart\Assert\Assert;
