+++ src/Component/Symfony/EventListener/WriteListener.php
@@ -14,8 +14,8 @@ declare(strict_types=1);
 namespace Sylius\Component\Resource\Symfony\EventListener;
 
 use Sylius\Component\Resource\Metadata\Operation\HttpOperationInitiatorInterface;
-use Sylius\Component\Resource\State\ProcessorInterface;
 use Sylius\Resource\Context\Initiator\RequestContextInitiatorInterface;
+use Sylius\Resource\State\ProcessorInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\ViewEvent;
