+++ src/Component/Symfony/Request/State/ApiResponder.php
@@ -17,10 +17,10 @@ use Sylius\Component\Resource\Metadata\CreateOperationInterface;
 use Sylius\Component\Resource\Metadata\DeleteOperationInterface;
 use Sylius\Component\Resource\Metadata\Operation;
 use Sylius\Component\Resource\Metadata\UpdateOperationInterface;
-use Sylius\Component\Resource\State\ResponderInterface;
 use Sylius\Component\Resource\Symfony\Response\HeadersInitiatorInterface;
 use Sylius\Resource\Context\Context;
 use Sylius\Resource\Context\Option\RequestOption;
+use Sylius\Resource\State\ResponderInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Webmozart\Assert\Assert;
