+++ src/Component/src/Doctrine/Common/State/RemoveProcessor.php
@@ -17,8 +17,8 @@ use Doctrine\Persistence\ManagerRegistry;
 use Doctrine\Persistence\ObjectManager as DoctrineObjectManager;
 use Sylius\Component\Resource\Metadata\Operation;
 use Sylius\Component\Resource\Reflection\ClassInfoTrait;
-use Sylius\Component\Resource\State\ProcessorInterface;
 use Sylius\Resource\Context\Context;
+use Sylius\Resource\State\ProcessorInterface;
 
 final class RemoveProcessor implements ProcessorInterface
 {
