+++ src/Component/src/State/Factory.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Resource\State;
+namespace Sylius\Resource\State;
 
 use Psr\Container\ContainerInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface as ResourceFactoryInterface;
