+++ src/Component/tests/spec/State/Processor/BulkAwareProcessorSpec.php
@@ -11,15 +11,15 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\State;
+namespace spec\Sylius\Resource\State\Processor;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Resource\Metadata\Api\Delete;
 use Sylius\Component\Resource\Metadata\BulkDelete;
-use Sylius\Component\Resource\State\BulkAwareProcessor;
-use Sylius\Component\Resource\State\ProcessorInterface;
 use Sylius\Component\Resource\Symfony\EventDispatcher\OperationEventDispatcherInterface;
 use Sylius\Resource\Context\Context;
+use Sylius\Resource\State\Processor\BulkAwareProcessor;
+use Sylius\Resource\State\ProcessorInterface;
 
 final class BulkAwareProcessorSpec extends ObjectBehavior
 {
