+++ src/Component/tests/spec/State/Processor/EventDispatcherBulkAwareProcessorSpec.php
@@ -11,16 +11,16 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\State;
+namespace spec\Sylius\Resource\State\Processor;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Resource\Metadata\BulkDelete;
-use Sylius\Component\Resource\State\EventDispatcherBulkAwareProcessor;
-use Sylius\Component\Resource\State\ProcessorInterface;
 use Sylius\Component\Resource\Symfony\EventDispatcher\OperationEvent;
 use Sylius\Component\Resource\Symfony\EventDispatcher\OperationEventDispatcherInterface;
 use Sylius\Component\Resource\Tests\Dummy\ProcessorWithCallable;
 use Sylius\Resource\Context\Context;
+use Sylius\Resource\State\Processor\EventDispatcherBulkAwareProcessor;
+use Sylius\Resource\State\ProcessorInterface;
 
 final class EventDispatcherBulkAwareProcessorSpec extends ObjectBehavior
 {
