+++ src/Component/tests/spec/State/Processor/EventDispatcherProcessorSpec.php
@@ -11,16 +11,16 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\State;
+namespace spec\Sylius\Resource\State\Processor;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Resource\Metadata\Create;
-use Sylius\Component\Resource\State\EventDispatcherProcessor;
-use Sylius\Component\Resource\State\ProcessorInterface;
 use Sylius\Component\Resource\Symfony\EventDispatcher\OperationEvent;
 use Sylius\Component\Resource\Symfony\EventDispatcher\OperationEventDispatcherInterface;
 use Sylius\Component\Resource\Symfony\EventDispatcher\OperationEventHandlerInterface;
 use Sylius\Resource\Context\Context;
+use Sylius\Resource\State\Processor\EventDispatcherProcessor;
+use Sylius\Resource\State\ProcessorInterface;
 use Symfony\Component\HttpFoundation\Response;
 
 final class EventDispatcherProcessorSpec extends ObjectBehavior
