+++ src/Component/tests/spec/State/ProcessorSpec.php
@@ -11,15 +11,15 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\State;
+namespace spec\Sylius\Resource\State;
 
 use PhpSpec\ObjectBehavior;
 use Psr\Container\ContainerInterface;
 use Sylius\Component\Resource\Metadata\Create;
-use Sylius\Component\Resource\State\Processor;
-use Sylius\Component\Resource\State\ProcessorInterface;
 use Sylius\Component\Resource\Tests\Dummy\ProcessorWithCallable;
 use Sylius\Resource\Context\Context;
+use Sylius\Resource\State\Processor;
+use Sylius\Resource\State\ProcessorInterface;
 
 final class ProcessorSpec extends ObjectBehavior
 {
@@ -73,7 +73,7 @@ final class ProcessorSpec extends ObjectBehavior
         $locator->has('\stdClass')->willReturn(true);
         $locator->get('\stdClass')->willReturn(new \stdClass());
 
-        $this->shouldThrow(new \InvalidArgumentException('Expected an instance of Sylius\Component\Resource\State\ProcessorInterface. Got: stdClass'))
+        $this->shouldThrow(new \InvalidArgumentException('Expected an instance of Sylius\Resource\State\ProcessorInterface. Got: stdClass'))
             ->during('process', [[], $operation, $context])
         ;
     }
