+++ src/Component/tests/spec/State/Provider/EventDispatcherProviderSpec.php
@@ -11,17 +11,17 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\State;
+namespace spec\Sylius\Resource\State\Provider;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Resource\Metadata\Create;
 use Sylius\Component\Resource\Metadata\Index;
 use Sylius\Component\Resource\Metadata\Show;
-use Sylius\Component\Resource\State\EventDispatcherProvider;
-use Sylius\Component\Resource\State\ProviderInterface;
 use Sylius\Component\Resource\Symfony\EventDispatcher\OperationEvent;
 use Sylius\Component\Resource\Symfony\EventDispatcher\OperationEventDispatcherInterface;
 use Sylius\Resource\Context\Context;
+use Sylius\Resource\State\Provider\EventDispatcherProvider;
+use Sylius\Resource\State\ProviderInterface;
 
 final class EventDispatcherProviderSpec extends ObjectBehavior
 {
