+++ src/Component/tests/spec/State/ResponderSpec.php
@@ -11,15 +11,15 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\State;
+namespace spec\Sylius\Resource\State;
 
 use PhpSpec\ObjectBehavior;
 use Psr\Container\ContainerInterface;
 use Sylius\Component\Resource\Metadata\Create;
-use Sylius\Component\Resource\State\Responder;
-use Sylius\Component\Resource\State\ResponderInterface;
 use Sylius\Component\Resource\Tests\Dummy\ResponderWithCallable;
 use Sylius\Resource\Context\Context;
+use Sylius\Resource\State\Responder;
+use Sylius\Resource\State\ResponderInterface;
 use Symfony\Component\HttpFoundation\Response;
 
 final class ResponderSpec extends ObjectBehavior
