+++ tests/Application/src/BoardGameBlog/Infrastructure/Sylius/State/Http/Processor/CreateBoardGameProcessor.php
@@ -19,8 +19,8 @@ use App\BoardGameBlog\Domain\ValueObject\BoardGameName;
 use App\BoardGameBlog\Infrastructure\Sylius\Resource\BoardGameResource;
 use App\Shared\Application\Command\CommandBusInterface;
 use Sylius\Component\Resource\Metadata\Operation;
-use Sylius\Component\Resource\State\ProcessorInterface;
 use Sylius\Resource\Context\Context;
+use Sylius\Resource\State\ProcessorInterface;
 use Webmozart\Assert\Assert;
 
 final class CreateBoardGameProcessor implements ProcessorInterface
