+++ tests/Application/src/BoardGameBlog/Infrastructure/Sylius/State/Http/Provider/BoardGameItemProvider.php
@@ -18,9 +18,9 @@ use App\BoardGameBlog\Domain\ValueObject\BoardGameId;
 use App\BoardGameBlog\Infrastructure\Sylius\Resource\BoardGameResource;
 use App\Shared\Application\Query\QueryBusInterface;
 use Sylius\Component\Resource\Metadata\Operation;
-use Sylius\Component\Resource\State\ProviderInterface;
 use Sylius\Resource\Context\Context;
 use Sylius\Resource\Context\Option\RequestOption;
+use Sylius\Resource\State\ProviderInterface;
 use Symfony\Component\Uid\Uuid;
 use Webmozart\Assert\Assert;
